feat(panda): Implement proxy support in panda live recorder
`{'result': False, 'message': '403 – Forbidden: Access is denied'}`
The updated code allows for the use of proxy servers when making requests from the panda live recorder. This is achieved by retrieving the 'proxy' key from the 'panda' dictionary in kwargs and using it in the HTTP request. Furthermore, sample configuration for the proxy server has also been added in the sample configuration toml file.

Signed-off-by: hldh214 <hldh214@gmail.com>