Update Douyin's danmu websocket URL and headers
The websocket URL for Douyin's danmu has been updated to reflect the changes in their streaming service. The URL now points to a different server, and some parameters in the URL have been changed including the webcast SDK version and room_id. The headers in the websocket connection have also been modified, removing the cookie and updating the user agent.

Signed-off-by: hldh214 <hldh214@gmail.com>